+++ packages/format-message/index.js
@@ -4,6 +4,7 @@
 var assign = require('object-assign')
 var parse = require('format-message-parse')
 var interpret = require('format-message-interpret')
+var plurals = require('format-message-interpret/plurals')
 var lookupClosestLocale = require('lookup-closest-locale')
 var formats = require('format-message-formats')
 
@@ -18,24 +19,24 @@ var generateId = function (pattern) { return pattern }
 var missingReplacement = null
 var missingTranslation = 'warning'
 
-module.exports = formatMessage
+module.exports = exports = formatMessage
 function formatMessage (msg, args, locale) {
   locale = locale || currentLocale
   var pattern = typeof msg === 'string' ? msg : msg.default
   var id = typeof msg === 'object' && msg.id || generateId(pattern)
   var key = locale + ':' + id
   var format = cache[key] ||
-    (cache[key] = generateFormat(locale, pattern, id))
+    (cache[key] = generateFormat(pattern, id, locale))
   if (typeof format === 'string') return format
   return format(args)
 }
 
-function generateFormat (locale, pattern, id) {
-  pattern = translate(locale, pattern, id)
+function generateFormat (pattern, id, locale) {
+  pattern = translate(pattern, id, locale)
   return interpret(locale, parse(pattern))
 }
 
-function translate (locale, pattern, id) {
+function translate (pattern, id, locale) {
   if (!translations) return pattern
 
   locale = lookupClosestLocale(locale, translations)
@@ -44,6 +45,9 @@ function translate (locale, pattern, id) {
   if (translated != null) return translated
 
   var replacement = missingReplacement || pattern
+  if (typeof replacement === 'function') {
+    replacement = replacement(pattern, id, locale) || pattern
+  }
   var message = 'Translation for "' + id + '" in "' + locale + '" is missing'
 
   if (missingTranslation === 'ignore') {
@@ -57,7 +61,7 @@ function translate (locale, pattern, id) {
   return replacement
 }
 
-formatMessage.setup = function setup (opt) {
+exports.setup = function setup (opt) {
   opt = opt || {}
   if (opt.locale) currentLocale = opt.locale
   if ('translations' in opt) translations = opt.translations
@@ -69,9 +73,18 @@ formatMessage.setup = function setup (opt) {
     if (opt.formats.date) assign(date, opt.formats.date)
     if (opt.formats.time) assign(time, opt.formats.time)
   }
+  return {
+    locale: currentLocale,
+    translations: translations,
+    generateId: generateId,
+    missingReplacement: missingReplacement,
+    missingTranslation: missingTranslation,
+    formats: { number: number, date: date, time: time }
+  }
 }
 
-formatMessage.number = function (locale, value, style) {
+exports.number = function (value, style, locale) {
+  locale = locale || currentLocale
   var options = number[style] || number.decimal
   if (typeof Intl === 'undefined') {
     return Number(value).toLocaleString(locale, options)
@@ -82,7 +95,8 @@ formatMessage.number = function (locale, value, style) {
   return format(value)
 }
 
-formatMessage.date = function (locale, value, style) {
+exports.date = function (value, style, locale) {
+  locale = locale || currentLocale
   var options = date[style] || date.medium
   if (typeof Intl === 'undefined') {
     return new Date(value).toLocaleDateString(locale, options)
@@ -93,7 +107,8 @@ formatMessage.date = function (locale, value, style) {
   return format(value)
 }
 
-formatMessage.time = function (locale, value, style) {
+exports.time = function (value, style, locale) {
+  locale = locale || currentLocale
   var options = time[style] || time.medium
   if (typeof Intl === 'undefined') {
     return new Date(value).toLocaleTimeString(locale, options)
@@ -103,3 +118,28 @@ formatMessage.time = function (locale, value, style) {
     (cache[locale] = new Intl.DateTimeFormat(locale, options).format)
   return format(value)
 }
+
+exports.select = function (value, options) {
+  return options[value] || options.other
+}
+
+function selectPlural (pluralType, value, offset, options, locale) {
+  if (typeof offset === 'object') { // offset is optional
+    locale = options
+    options = offset
+    offset = 0
+  }
+
+  var closest = lookupClosestLocale(locale || currentLocale, plurals)
+  var plural = plurals[closest][pluralType]
+  if (!plural) return options.other
+
+  return (
+    options['=' + +value] ||
+    options[plural(value - offset)] ||
+    options.other
+  )
+}
+
+exports.plural = selectPlural.bind(null, 'cardinal')
+exports.selectordinal = selectPlural.bind(null, 'ordinal')